requirements.txt
# API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Azure Services
azure-ai-formrecognizer==3.3.3
azure-core==1.29.7
azure-identity==1.15.0

# Database
pymongo==4.6.1
motor==3.3.2

# Supabase (commenté - non nécessaire, on construit les URLs manuellement)
# supabase==2.3.4

# HTTP & File handling
httpx==0.27.2
aiohttp==3.11.11
requests==2.31.0
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
tenacity==8.2.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-httpx==0.30.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
